attendee controller crud added, swagger ui added

+++ app/src/main/java/com/codesherpa/beerdispenser/app/repositories/AttendeeRepository.java
@@ -3,5 +3,5 @@ package com.codesherpa.beerdispenser.app.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.codesherpa.beerdispenser.app.models.Attendee;
-public interface AttendeeRepository extends JpaRepository<Attendee, Integer> {
+public interface AttendeeRepository extends JpaRepository<Attendee, Long> {
 }
